+++ src/app/component-library/json-dialog/json-dialog.module.ts
@@ -2,6 +2,7 @@ import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {MatButtonModule} from '@angular/material/button';
+import {MatCardModule} from '@angular/material/card';
 import {MatCheckboxModule} from '@angular/material/checkbox';
 import {MatDialogModule} from '@angular/material/dialog';
 import {MatIconModule} from '@angular/material/icon';
@@ -10,6 +11,7 @@ import {NgxJsonViewerModule} from 'ngx-json-viewer';
 import {ConfirmationDialogModule} from '../confirmation-dialog/confirmation-dialog.module';
 import {PipesAndDirectivesModule} from '../pipes-and-directives/pipes-and-directives.module';
 import {JsonDialogComponent} from './json-dialog/json-dialog.component';
+import {JsonDialogService} from './json-dialog/json-dialog.service';
 
 @NgModule({
   imports: [
@@ -19,6 +21,7 @@ import {JsonDialogComponent} from './json-dialog/json-dialog.component';
 
     // Angular Material
     MatButtonModule,
+    MatCardModule,
     MatCheckboxModule,
     MatDialogModule,
     MatIconModule,
@@ -32,6 +35,7 @@ import {JsonDialogComponent} from './json-dialog/json-dialog.component';
     ConfirmationDialogModule,
   ],
   declarations: [JsonDialogComponent],
+  providers: [JsonDialogService],
   exports: [JsonDialogComponent],
 })
 export class JsonDialogModule {}