+++ src/app/routes/connector-ui/catalog-browser-page/catalog-browser-page/catalog-browser-page.component.ts
@@ -8,10 +8,9 @@ import {
   distinctUntilChanged,
   sampleTime,
 } from 'rxjs';
-import {map} from 'rxjs/operators';
+import {filter, map} from 'rxjs/operators';
 import {AssetDetailDialogDataService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog-data.service';
-import {AssetDetailDialogResult} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog-result';
-import {AssetDetailDialogComponent} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.component';
+import {AssetDetailDialogService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.service';
 import {CatalogApiUrlService} from '../../../../core/services/api/catalog-api-url.service';
 import {ContractOffer} from '../../../../core/services/models/contract-offer';
 import {value$} from '../../../../core/utils/form-group-utils';
@@ -35,6 +34,7 @@ export class CatalogBrowserPageComponent implements OnInit, OnDestroy {
 
   constructor(
     private assetDetailDialogDataService: AssetDetailDialogDataService,
+    private assetDetailDialogService: AssetDetailDialogService,
     private catalogBrowserPageService: CatalogBrowserPageService,
     private catalogApiUrlService: CatalogApiUrlService,
     private matDialog: MatDialog,
@@ -56,12 +56,10 @@ export class CatalogBrowserPageComponent implements OnInit, OnDestroy {
   onContractOfferClick(contractOffer: ContractOffer) {
     const data =
       this.assetDetailDialogDataService.contractOfferDetails(contractOffer);
-    const ref = this.matDialog.open(AssetDetailDialogComponent, {data});
-    ref.afterClosed().subscribe((result: AssetDetailDialogResult) => {
-      if (result?.refreshList) {
-        this.fetch$.next(null);
-      }
-    });
+    this.assetDetailDialogService
+      .open(data, this.ngOnDestroy$)
+      .pipe(filter((it) => !!it?.refreshList))
+      .subscribe(() => this.fetch$.next(null));
   }
 
   onShowFetchDetails() {