+++ src/app/routes/connector-ui/contract-definition-page/asset-select/asset-select.component.ts
@@ -1,15 +1,15 @@
-import {Component, Input} from '@angular/core';
+import {Component, Input, OnDestroy} from '@angular/core';
 import {FormControl} from '@angular/forms';
-import {MatDialog} from '@angular/material/dialog';
+import {Subject} from 'rxjs';
 import {AssetDetailDialogDataService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog-data.service';
-import {AssetDetailDialogComponent} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.component';
+import {AssetDetailDialogService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.service';
 import {Asset} from '../../../../core/services/models/asset';
 
 @Component({
   selector: 'asset-select',
   templateUrl: './asset-select.component.html',
 })
-export class AssetSelectComponent {
+export class AssetSelectComponent implements OnDestroy {
   @Input()
   label!: string;
 
@@ -21,14 +21,18 @@ export class AssetSelectComponent {
 
   constructor(
     private assetDetailDialogDataService: AssetDetailDialogDataService,
-    private matDialog: MatDialog,
+    private assetDetailDialogService: AssetDetailDialogService,
   ) {}
 
   onAssetClick(asset: Asset) {
     const data = this.assetDetailDialogDataService.assetDetails(asset, false);
-    this.matDialog.open(AssetDetailDialogComponent, {
-      data,
-      maxHeight: '90vh',
-    });
+    this.assetDetailDialogService.open(data, this.ngOnDestroy$);
+  }
+
+  ngOnDestroy$ = new Subject();
+
+  ngOnDestroy() {
+    this.ngOnDestroy$.next(null);
+    this.ngOnDestroy$.complete();
   }
 }