+++ src/app/routes/connector-ui/dashboard-page/dashboard-page/dashboard-page-data.service.ts
@@ -1,6 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Observable, combineLatest, merge, of, scan} from 'rxjs';
 import {map} from 'rxjs/operators';
+import {TransferHistoryEntry} from '@sovity.de/edc-client';
 import {CatalogApiUrlService} from '../../../../core/services/api/catalog-api-url.service';
 import {ContractOfferService} from '../../../../core/services/api/contract-offer.service';
 import {EdcApiService} from '../../../../core/services/api/edc-api.service';
@@ -8,18 +9,15 @@ import {
   ContractAgreementService,
   ContractDefinitionService,
   PolicyService,
-  TransferProcessDto,
-  TransferProcessService,
 } from '../../../../core/services/api/legacy-managent-api-client';
 import {
   ConnectorInfoPropertyGridGroupBuilder
 } from '../../../../core/services/connector-info-property-grid-group-builder';
 import {LastCommitInfoService} from '../../../../core/services/last-commit-info.service';
 import {Fetched} from '../../../../core/services/models/fetched';
-import {TransferProcessStates} from '../../../../core/services/models/transfer-process-states';
-import {TransferProcessUtils} from '../../../../core/services/transfer-process-utils';
 import {DonutChartData} from '../dashboard-donut-chart/donut-chart-data';
 import {DashboardPageData, defaultDashboardData} from './dashboard-page-data';
+
 
 @Injectable({providedIn: 'root'})
 export class DashboardPageDataService {
@@ -30,11 +28,10 @@ export class DashboardPageDataService {
     private contractAgreementService: ContractAgreementService,
     private policyService: PolicyService,
     private catalogApiUrlService: CatalogApiUrlService,
-    private transferProcessService: TransferProcessService,
-    private transferProcessUtils: TransferProcessUtils,
     private lastCommitInfoService: LastCommitInfoService,
     private connectorInfoPropertyGridGroupBuilder: ConnectorInfoPropertyGridGroupBuilder,
-  ) {}
+  ) {
+  }
 
   /**
    * Fetch {@link DashboardPageData}.
@@ -120,37 +117,32 @@ export class DashboardPageDataService {
   }
 
   private transferProcessKpis(): Observable<Partial<DashboardPageData>> {
-    return this.transferProcessService
-      .getAllTransferProcesses(0, 10_000_000)
+    return this.edcApiService.getTransferHistoryPage()
       .pipe(
         Fetched.wrap({
           failureMessage: 'Failed fetching transfer processes.',
         }),
         map((transferData) => ({
           incomingTransfersChart: transferData.map((it) =>
-            this.buildTransferChart(it, 'incoming'),
+            this.buildTransferChart(it.transferEntries, 'CONSUMING'),
           ),
           outgoingTransfersChart: transferData.map((it) =>
-            this.buildTransferChart(it, 'outgoing'),
+            this.buildTransferChart(it.transferEntries, 'PROVIDING'),
           ),
         })),
       );
   }
 
   private buildTransferChart(
-    transfers: TransferProcessDto[],
-    direction: 'incoming' | 'outgoing',
+    transfers: TransferHistoryEntry[],
+    direction: 'CONSUMING' | 'PROVIDING',
   ): DonutChartData {
     const filteredTransfers =
-      direction === 'incoming'
-        ? transfers.filter((it) => this.transferProcessUtils.isIncoming(it))
-        : transfers.filter((it) => this.transferProcessUtils.isOutgoing(it));
+      direction === 'CONSUMING'
+        ? transfers.filter((it) => it.direction === 'CONSUMING')
+        : transfers.filter((it) => it.direction === 'PROVIDING');
 
-    // Use the keys of the TransferProcessesStates Enum as order
-    const order = Object.keys(TransferProcessStates);
-    const states = [...new Set(filteredTransfers.map((it) => it.state))].sort(
-      (a, b) => order.indexOf(a) - order.indexOf(b),
-    );
+    const states = [...new Set(filteredTransfers.sort((a, b) => a.state.code - b.state.code).map((it) => it.state.name))]
 
     const colorsByState = new Map<string, string>();
     colorsByState.set('IN_PROGRESS', '#7eb0d5');
@@ -159,7 +151,7 @@ export class DashboardPageDataService {
     const defaultColor = '#bd7ebe';
 
     const amountsByState = states.map(
-      (state) => filteredTransfers.filter((it) => it.state === state).length,
+      (state) => filteredTransfers.filter((it) => it.state.name === state).length,
     );
 
     return {