+++ src/app/routes/connector-ui/transfer-history-page/transfer-history-page.module.ts
@@ -2,10 +2,16 @@ import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {MatButtonModule} from '@angular/material/button';
+import {MatCardModule} from '@angular/material/card';
+import {MatDialogModule} from '@angular/material/dialog';
 import {MatIconModule} from '@angular/material/icon';
 import {MatPaginatorModule} from '@angular/material/paginator';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 import {MatTableModule} from '@angular/material/table';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {RouterModule} from '@angular/router';
+import {CatalogModule} from '../../../component-library/catalog/catalog.module';
+import {JsonDialogModule} from '../../../component-library/json-dialog/json-dialog.module';
 import {PipesAndDirectivesModule} from '../../../component-library/pipes-and-directives/pipes-and-directives.module';
 import {UiElementsModule} from '../../../component-library/ui-elements/ui-elements.module';
 import {TransferHistoryPageComponent} from './transfer-history-page/transfer-history-page.component';
@@ -20,11 +26,17 @@ import {TransferHistoryPageComponent} from './transfer-history-page/transfer-his
 
     // Angular Material
     MatButtonModule,
-    MatTableModule,
-    MatPaginatorModule,
+    MatCardModule,
+    MatDialogModule,
     MatIconModule,
+    MatPaginatorModule,
+    MatProgressSpinnerModule,
+    MatTableModule,
+    MatTooltipModule,
 
     // EDC UI Modules
+    CatalogModule,
+    JsonDialogModule,
     PipesAndDirectivesModule,
     UiElementsModule,
   ],